cspell and clippy

+++ src/roinput.rs
@@ -112,7 +112,7 @@ impl ROInput {
   pub fn append_scalar(mut self, s: Fq) -> Self {
     // mina scalars are 255 bytes
     let bytes = s.to_bytes();
-    let bits = &bytes.as_bits::<Lsb0>()[.. <Fq as PrimeField>::size_in_bits() as usize];
+    let bits = &bytes.as_bits::<Lsb0>()[.. <Fq as PrimeField>::size_in_bits()];
     self.bits.extend(bits);
     self
   }
@@ -148,7 +148,7 @@ impl ROInput {
   /// Serialize random oracle input to bytes
   pub fn to_bytes(&self) -> Vec<u8> {
     let mut bits: BitVec<u8> = self.fields.iter().fold(BitVec::new(), |mut acc, fe| {
-      acc.extend_from_bitslice(&fe.to_bytes().as_bits::<Lsb0>()[.. <Fp as PrimeField>::size_in_bits() as usize]);
+      acc.extend_from_bitslice(&fe.to_bytes().as_bits::<Lsb0>()[.. <Fp as PrimeField>::size_in_bits()]);
 
       acc
     });
@@ -162,32 +162,31 @@ impl ROInput {
   pub fn to_fields(&self) -> Vec<Fp> {
     let mut fields: Vec<Fp> = self.fields.clone();
 
-    let bits_as_fields =
-      self.bits.chunks(<Fp as PrimeField>::size_in_bits() as usize - 1).fold(vec![], |mut acc, chunk| {
-        // Workaround: chunk.clone() does not appear to respect
-        // the chunk's boundaries when it's not byte-aligned.
-        //
-        // That is,
-        //
-        //   let mut bv = chunk.clone().to_bitvec();
-        //   bv.resize(B::size_in_bits(), false);
-        //   fields.push(B::from_bytes(bv.into()));
-        //
-        // doesn't work.
-        //
-        // Instead we must do
+    let bits_as_fields = self.bits.chunks(<Fp as PrimeField>::size_in_bits() - 1).fold(vec![], |mut acc, chunk| {
+      // Workaround: chunk.clone() does not appear to respect
+      // the chunk's boundaries when it's not byte-aligned.
+      //
+      // That is,
+      //
+      //   let mut bv = chunk.clone().to_bitvec();
+      //   bv.resize(B::size_in_bits(), false);
+      //   fields.push(B::from_bytes(bv.into()));
+      //
+      // doesn't work.
+      //
+      // Instead we must do
 
-        let mut bv = BitVec::<u8>::new();
-        bv.resize(chunk.len(), false);
-        bv.clone_from_bitslice(chunk);
+      let mut bv = BitVec::<u8>::new();
+      bv.resize(chunk.len(), false);
+      bv.clone_from_bitslice(chunk);
 
-        // extend to the size of a field;
-        bv.resize(<Fp as PrimeField>::size_in_bits() as usize, false);
+      // extend to the size of a field;
+      bv.resize(<Fp as PrimeField>::size_in_bits(), false);
 
-        acc.push(Fp::from_bytes(&bv.into_vec()).expect("failed to create base field element"));
+      acc.push(Fp::from_bytes(&bv.into_vec()).expect("failed to create base field element"));
 
-        acc
-      });
+      acc
+    });
 
     fields.extend(bits_as_fields);
